producer/requirements.txt
kafka-python
pandas
python-dotenv